// src/app/models/election.ts
export class Election {
    prop0: string;
    prop1: string;
    date: string;
    unix: number;
    register: string;
    proposal0: string;
    proposal1: string;
    proposal0Count: number;
    proposal1Count: number;
    show: boolean;
    blockTimestamp: string;
    getDate: string;
    revokeAddr: string;
}